fix: move prettier in app
