+++ src/pages/Editor/index.tsx
@@ -15,7 +15,7 @@ import { selectIsUser } from '../../app/store/slices/authSlices';
 import { HeadersEditor, QueryEditor, VariableEditor } from '../../shared/ui';
 import { getSchema } from '../../shared/utils/getSchema';
 import { getAPI } from '../../shared/utils/getApi';
-import { convertToPrettier } from '../../shared/utils/prettierApp';
+import { convertToPrettier } from '../../app/prettier/prettierApp';
 import { AppPrettier } from '../../app/assets/icons/AppPrettier';
 import styles from './Editor.module.scss';
 import { useAppSelector } from '../../app/store/hooks/hooks';
